requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloud-tpu-client==0.5
cloud-tpu-profiler==1.15.0rc1
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
distro==1.0.1
entrypoints==0.3
filelock==3.0.12
fire==0.2.1
future==0.18.2
gast==0.2.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-compute-engine==20191210.0
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
httplib2==0.17.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
ipykernel==5.1.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0
Pygments==2.5.2
pygobject==3.22.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-apt==1.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rsa==4.0
scipy==1.4.1
seaborn==0.9.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-datasets==2.1.0
tensorflow-estimator==1.15.1
tensorflow-gan==0.0.0.dev0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.1
tensorflow-probability==0.8.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==20.0.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
zipp==1.2.0
